add fb connect

+++ views.py
@@ -33,6 +33,58 @@ def showLogin():
     login_session['state'] = state
     return render_template('login.html', STATE=state)
 
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s" % access_token
+
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+            'web']['app_id']
+    app_secret = json.loads(
+            open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+                     app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+    user_info_url = "https://graph.facebook.com/v2.8/me"
+
+    token = result.split(',')[0].split(':')[1].replace('"','')
+    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data['email']
+    login_session['email'] = data['email']
+    login_session['facebook_id'] = data['id']
+    login_session['access_token'] = token
+    #get pic
+    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+    login_session['picture'] = data["data"]["url"]
+
+    user_id = getUserId(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
+
 @app.route('/addcookbook/', methods=['GET', 'POST'])
 def addCookBook():
     session = DBSession()
@@ -70,8 +122,8 @@ def addCookBook():
 @app.route('/deletebook/<int:book_id>/', methods=['GET', 'POST'])
 def deleteCookBook(book_id):
     session = DBSession()
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     bookToDelete = session.query(Book).filter_by(id = book_id).one()
     if request.method == 'POST':
         session.delete(bookToDelete)
@@ -83,8 +135,8 @@ def deleteCookBook(book_id):
 @app.route('/cookbook/edit/<int:book_id>/', methods=['GET', 'POST'])
 def editCookBook(book_id):
     session = DBSession()
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     bookToEdit = session.query(Book).filter_by(id = book_id).one()
     bookToEditAuthor = session.query(Author).filter_by(id = bookToEdit.author_id).first()
     categories = session.query(Category).all()
